+++ src/main/java/ecommerce/Product/ExpirableProduct.java
@@ -24,11 +24,11 @@ public class ExpirableProduct implements Product, Expirable {
 		return product.getQuantity();
 	}
 
-	public int getPrice() {
+	public double getPrice() {
 		return product.getPrice();
 	}
 
-	public int getTotalPrice() {
+	public double getTotalPrice() {
 		return product.getTotalPrice();
 	}
